[SPARC] Make calls to function with big return values work
Implement CanLowerReturn and associated CallingConv changes for SPARC/SPARC64.

In particular, for SPARC64 there's new RetCC_Sparc64_* functions that handles
the return case of the calling convention. It uses the same analysis as
CC_Sparc64_* family of funtions, but fails if the return value doesn't fit into
the return registers.

This makes calls to functions with big return values converted to an sret
function as expected, instead of crashing LLVM.